* remove runner

* fix custom images

* use old hook that seems to work 100%

* update premake

* og & custom colors

* add rocket jump dvar

* poosh

* add r_drawFxInfo

* reflection probe fix and exit

* fix http crash

* fix build

* allow using unauthorized clantags

* Update gameplay.cpp

* read all arenas instead of just one

* add r_colormode

* Fix probe exit check to use enabled state

The reflection probe generation exit now checks if the dvar is enabled before calling game::Quit(), preventing unintended exits when the dvar is not active.

* add imgui techset view

* add imgui rawfile view

* imgui techset view changes

* update structs

* fix steam proxy app id

* update database

* add lookupsoundlength gsc function

* add imgui shader view

* add d3d11 debug stuff

* some changes

* print current shaders during d3d11 error

* small fix

* add markdangerousnodes/intrigger funcs (gsc)

* fix cursor not working for .menu files

* add jump pathnodes

* add 3D drawing GSC helpers

* remove annoying override asset print

* change DEBUG to _DEBUG

* revise readme and remove useless premake arguments

* remove readme oopsie

* reload omnvars on loadmod

* optimizations & disable system_check temporarily

* add CSV path node loading

* allow ttf overriding

* remove unsafe lua funcs

* load ui scripts from rawfiles

* update ui scripts

* Update fonts.cpp

* reset fonts before vid restart (fix crash)

* revert workflow to h1-mod.exe only

* add lua zoneexists func

* Add debug GUI for viewing server config strings

* fix conststrings gui

* update map_patches

iw6 patches mostly

* small fix

* fix generate path

* menu overriding

* fix

* custom menu fonts

* custom hudelem fonts

* ddl overriding

* lua bitwise funcs

* fix crash

* lua script read changes

* remove mod/mod_pre_gfx checks

* let moveSpeedScale be used in GSC

* disable vid_restart for mod loading

* add gametype to lui server data

* fix infinite loop with > 255 weapons

* Update fastfiles.cpp

* small fix

* handle custom vlobby maps

* fix rocket jump hook

* remove rocket jump hook

* better rocket jump scale hook

* smal fix

* dont wait for vlobby path ff

* allow rootmotion in animscripts

* fix anim rate being set to inf

* add r_enableWarnings

* add statetimers to weapon gui

* chore: use correct links in menus (#1055)

* refactor: update checksum and files used on system_check component (#1054)

* update & sync readme [skip ci]

* W copy and paster

* add stdout

* add xpartygogametype

* remove voice chat settings in SP

* feat: MW2 player mechanics

this code originated from h2m-mod when I worked on it, and was improved by Patoke, REVLIIS, and other community members. thank you :)

* set genericMaterialData in r_adddobjtoscene

* fix gentity_s struct

* better imgui debug entity bounds

* fix spawn model button on imgui menu

* some imgui debug stuff

* add netconststrings imgui menu

* better configstring list

* only use usermaps if it exists

* handle fragmented responses with EOT

closes #851

* fix "out of memory" error for SP

taken from 3e6f54b9dff65cfcf2dd37beba60902aa124a110


* fix SP audio menu crash and restore subtitles option

taken from 34efbcbb2a0207fc59909d4800ac69ffbaff3e80


* remove unsafe LUI functions only on MP

* use h1_mp64_ship.exe module for AMD vendor fixes (#1059)

* spoof as h1_mp64_ship.exe for AMD vendor fixes

* remove comment

---------

+++ src/common/utils/http.cpp
@@ -85,7 +85,7 @@ namespace utils::http
 		}
 
 		const auto code = curl_easy_perform(curl);
-		unsigned int response_code{};
+		long response_code{};
 		curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &response_code);
 
 		if (code == CURLE_OK)
